데이터 스토어 객체를 만들어서 별도로 분리
- 데이터 스토어 클래스를 만들어 분리.
- 스테이트 플로를 만들어 쿼리를 저장하게 하고 변경시 데이터 스토어를 다시 생성하게 함.
