+++ app/src/main/java/io/github/dalinaum/m_issues/viewModel/GithubViewModel.kt
@@ -1,13 +1,16 @@
 package io.github.dalinaum.m_issues.viewModel
 
-import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import androidx.paging.*
 import dagger.hilt.android.lifecycle.HiltViewModel
 import io.github.dalinaum.m_issues.api.GithubService
 import io.github.dalinaum.m_issues.data.Item
+import io.github.dalinaum.m_issues.dataSource.GithubDataSource
+import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.flatMapLatest
 import javax.inject.Inject
 
 
@@ -15,8 +18,7 @@ import javax.inject.Inject
 class GithubViewModel @Inject constructor(
     private val githubService: GithubService
 ) : ViewModel() {
-    val searchedRepositories = getRepositoriesFlow("D").cachedIn(viewModelScope)
-
+    private val currentQuery = MutableStateFlow("D")
     private val pageSize = 30
 
     private fun getRepositoriesFlow(query: String): Flow<PagingData<Item>> = Pager(
@@ -27,33 +29,15 @@ class GithubViewModel @Inject constructor(
         ),
         initialKey = 1
     ) {
-        object : PagingSource<Int, Item>() {
-            override fun getRefreshKey(state: PagingState<Int, Item>): Int? {
-                return state.anchorPosition?.let { anchorPosition ->
-                    state.closestPageToPosition(anchorPosition)?.prevKey?.plus(1)
-                        ?: state.closestPageToPosition(anchorPosition)?.nextKey?.minus(1)
-                }
-            }
-
-            override suspend fun load(params: LoadParams<Int>): LoadResult<Int, Item> {
-                return try {
-                    val response = githubService.searchRepositories(
-                        searchQuery = query,
-                        page = params.key,
-                        perPage = pageSize
-                    )
-
-                    LoadResult.Page(
-                        data = response.items,
-                        prevKey = if (params.key == 1) null else params.key?.minus(1),
-                        nextKey = if (response.items.isEmpty()) null else params.key?.plus(1)
-                    )
-                } catch (e: Exception) {
-                    Log.e("EEE", "error: $e")
-                    e.printStackTrace()
-                    LoadResult.Error(e)
-                }
-            }
-        }
+        GithubDataSource(githubService, query, pageSize)
     }.flow
+
+    @OptIn(ExperimentalCoroutinesApi::class)
+    val searchedRepositories: Flow<PagingData<Item>> = currentQuery.flatMapLatest {
+        getRepositoriesFlow(it).cachedIn(viewModelScope)
+    }
+
+    fun setQuery(query: String) {
+        currentQuery.value = query
+    }
 }
